requirements.txt
gradio==4.44.0
pandas==2.2.2
python-dotenv==1.0.1
requests==2.32.3